+++ src/utility.h
@@ -432,10 +432,10 @@ MatchType str_n_nosp_cmp_begin(string arg1, string arg2);
 char *str_nospace(const char *stri);
 char *str_dup(const char *str);
 char *str_dup0(const char *str);
-void logentry(QString str, int god_level = 0, LogChannels type = LogChannels::LOG_MISC, Character *vict = nullptr);
+void logentry(QString str, uint64_t god_level = 0, LogChannels type = LogChannels::LOG_MISC, Character *vict = nullptr);
 void logf(int level, LogChannels type, const char *arg, ...);
 void logf(int level, LogChannels type, QString arg);
-int send_to_gods(QString message, int god_level, LogChannels type);
+int send_to_gods(QString message, uint64_t god_level, LogChannels type);
 
 void sprintbit(uint value[], const char *names[], char *result);
 std::string sprintbit(uint value[], const char *names[]);
